fix column order random bug in get_table (#24)

+++ greenplumpython/core/gpapply.py
@@ -65,7 +65,7 @@ def gpApply(dataframe, py_func, db, output, clear_existing = True, runtime_id = 
     if db == None:
         raise ValueError("No database connection provided")
 
-    if clear_existing:
+    if clear_existing and output.name != None and output.name != "":
         drop_table_sql = "drop table if exists %s;" % output.name
         db.execute(drop_table_sql)
     db.execute(create_type_sql)
